ingestion/requirements.txt
requests
google-cloud-storage
python-dotenv
google-cloud-aiplatform